Add files via upload
In the modern Era of technology, Chatbots is the next big thing in the era of conversational services. 
A chatbot or conversational agent is a software that can communicate with a human. This is 
becoming a big trend because it makes interacting with technology more natural, comfortable and 
user-friendly. People can find information or complete tasks by having a conversation, such as 
texting a friend, is a more efficient way rather than clicking buttons or filling out forms. Chatbots 
can be built using a variety of technologies, including Natural  
Language Processing and Application Programming Interface, to communicate via text or  voice. 
For example, It can be difficult to find specific information on a college's website without knowing 
where to look. For someone who is not a student or employee, it becomes more challenging to 
obtain information. A college inquiry chat bot is the result of these issues. It is a quick, easy-to
use, and educational widget designed to improve user experience on college websites and give 
users useful information.